server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.6
frozenlist==1.8.0
google-ai-generativelanguage==0.6.10
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
gotrue==2.9.1
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
limits==5.8.0
multidict==6.7.1
openai==2.21.0
packaging==26.0
postgrest==0.15.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.25.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
realtime==1.0.6
requests==2.32.5
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
storage3==0.7.7
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
tenacity==9.1.4
tqdm==4.67.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.34.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.1.1
yarl==1.22.0
youtube-transcript-api==0.6.3
